test(qtensor): isolate linear dispatch tests
